chore: move api call to useWeatherInfo hook
