+++ src/service/CityInfoProvider.ts
@@ -1,0 +1,24 @@
+interface CityProvider {
+  getCityIds(): number[];
+}
+
+class DefaultCityProvider implements CityProvider {
+  getCityIds() {
+    const cityIds = [
+      703448, // Kyiv, UA
+      692194, // Sumy, UA
+      756135, // Warsaw, PL
+      3081368, // Wrocław, PL
+      3067696, // Prague, CZ
+      3077916, // České Budějovice, CZ
+      2950159, // Berlin, DE
+      2867714, // Munich, DE
+      3247449, // Aachen, DE
+      5815135, // Washington, US
+      5128581, // New York City, US
+    ];
+    return cityIds;
+  }
+}
+
+export default DefaultCityProvider;
